sync lib with k8s pr

+++ lib/charms/mysql/v0/mysql.py
@@ -135,7 +135,7 @@ class MySQLDeleteUsersForUnitError(Error):
 
 
 class MySQLDeleteUsersForRelationError(Error):
-    """Exception raised when there is an issue deleting a user for a relation."""
+    """Exception raised when there is an issue deleting users for a relation."""
 
 
 class MySQLConfigureInstanceError(Error):
@@ -569,6 +569,8 @@ class MySQLBase(ABC):
         try:
             user = f"relation-{str(relation_id)}"
             primary_address = self.get_cluster_primary_address()
+            if not primary_address:
+                raise MySQLDeleteUsersForRelationError("Unable to query cluster primary address")
             drop_users_command = (
                 f"shell.connect('{self.server_config_user}:{self.server_config_password}@{primary_address}')",
                 f"session.run_sql(\"DROP USER IF EXISTS '{user}'@'%';\")",
